Redistribute classes into respective packages
