+++ src/main/java/duke/Parser.java
@@ -1,3 +1,6 @@
+package duke;
+
+import duke.command.*;
 import java.util.ArrayList;
 import java.util.Scanner;
