+++ src/main/java/duke/Ui.java
@@ -1,7 +1,9 @@
+package duke;
+
 import java.util.Scanner;
 
 public class Ui {
-    private static final String WELCOME_MESSAGE = "Hello, I'm Duke!\nWhat can I do for you?";
+    private static final String WELCOME_MESSAGE = "Hello, I'm duke.Duke!\nWhat can I do for you?";
     private static final String LINE_DIVIDER = "______________________________________";
 
     public Ui() {
